requirements.txt
streamlit
openai
pandas
pillow
pdf2image